requirements.txt
aiogram==3.13.1
httpx==0.27.2
python-dotenv==1.0.1
aiosqlite==0.20.0
apscheduler==3.10.4
pydantic==2.8.2

